Edited cleanup logic.

+++ lib/cleanup.py
@@ -1,9 +1,17 @@
 import os
 
-file = "class_count.txt"
-os.remove(file)
-print("%s has been removed successfully" % file)
 
-file = "mail_count.txt"
-os.remove(file)
-print("%s has been removed successfully" % file)
+def cleanup(root_folder):
+
+    root_folder = os.path.normcase(root_folder)
+    count_files = [os.path.join(root_folder, "lib", "class_count.txt"),
+                   os.path.join(root_folder, "lib", "mail_count.txt"),
+                   os.path.join(root_folder, "class_count.txt"),
+                   os.path.join(root_folder, "mail_count.txt")]
+
+    print()
+    for file in count_files:
+        if os.path.isfile(file):
+            os.remove(file)
+
+    print("Cleanup Successful.")
